layout/blspec.py: also clean-up UKI's with 'gentoo' prefix
With version 8 installkernel-gentoo gains an optional plugin to install UKIs
to the usual location at ${ESP}/EFI/Linux (i.e. BLS type 2 layout). This
is useful for boot loaders such as systemd-boot that auto discover UKIs at
this location, as well as for EFI stub booting where it is required to have
the UKI on the ESP (which is not necessarily mounted at /boot).

Because the entry-token that is the usual prefix for these UKIs is specific
to systemd and kernel-install (i.e. installkernel-systemd), installkernel-gentoo
prefixes the UKIs with 'gentoo' instead. The changes here ensure that eclean-
kernel recognizes these UKIs as ours and cleans them up as well.

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>
Closes: https://github.com/projg2/eclean-kernel/pull/45
Signed-off-by: Michał Górny <mgorny@gentoo.org>