bootloader: consider GRUB_CFG as location for grub config
sys-kernel/installkernel[grub] supports setting GRUB_CFG in the environment to
update a grub.cfg in a non-standard location. Specifically this is useful with
sys-boot/grub[secureboot] where the built grub efi executable loads the
grub.cfg from the same directory it is in.

Recognize the same environment variable here so we don't accidentally update
the wrong config.

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>
Closes: https://github.com/projg2/eclean-kernel/pull/57
Signed-off-by: Michał Górny <mgorny@gentoo.org>